moved 'ImageCommentLoaderPresentationAdapter' to it's own file
